requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.3.0
flask>=2.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0